+++ src/CssRgbColor/CssRgbColor.ts
@@ -44,7 +44,8 @@ import type { CssRgbColorChannelsData } from "./CssRgbColorChannelsData.type";
 import type { CssRgbColorChannelsTuple } from "./CssRgbColorChannelsTuple.type";
 import { DEFAULT_DATA_PATH, THROW_THE_ERROR, type FunctionalOption, type DataGuaranteeOptions } from "@safelytyped/core-types";
 import type { CssHslColorData } from "../CssHslColor/CssHslColorData.type";
-import { makeCssRgbColorData, type CssHwbColorData } from "@safelytyped/css-color";
+import type { CssHwbColorData } from "../CssHwbColor/CssHwbColorData.type";
+import { makeCssRgbColorData } from "./makeCssRgbColorData";
 
 /**
  * CssRgbColor represents a {@link CssColor} that was defined using the