+++ src/index.ts
@@ -151,10 +151,13 @@ export { hasClearContrast } from "./inspectors/hasClearContrast";
 export { isDark } from "./inspectors/isDark";
 export { isDull } from "./inspectors/isDull";
 export { isLight } from "./inspectors/isLight";
+export { isMidtone } from "./inspectors/isMidtone";
+export { isMonochrome } from "./inspectors/isMonochome";
 export { lightModeContrastRatio } from "./inspectors/lightModeContrastRatio";
 export { luma } from "./inspectors/luma";
 export { relativeLuminance } from "./inspectors/relativeLuminance";
 export { type Shade, shade } from "./inspectors/shade";
+export { type Tonality, tonality } from "./inspectors/tonality";
 
 export { type Hue, hues } from "./inspectors/hues/hues";
 export { isBlackHue } from "./inspectors/hues/isBlackHue";